+++ ora.go
@@ -25,7 +25,7 @@ const (
 
 	// The driver version sent to an Oracle server and visible in
 	// V$SESSION_CONNECT_INFO or GV$SESSION_CONNECT_INFO.
-	Version string = "v3.7.0"
+	Version string = "v3.7.1"
 )
 
 var _drv *Drv